+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v1.8.3
 
+### Fixed
+
 - Fix the parsing of `argv` when handling the `wp-cli` command.
+- Adjusted the way that `nunomaduro/collision` is loaded in PHPUnit.
 
 ## v1.8.2